[wyah]: Split files so that things are more focused
