+++ write-you-a-haskell/src/infer.ts
@@ -1,156 +1,11 @@
-import { Map, Set } from "immutable";
+import { Map } from "immutable";
 
-import {
-  InfiniteType,
-  UnboundVariable,
-  UnificationFail,
-  UnificationMismatch,
-} from "./errors";
-import {
-  Type,
-  TVar,
-  Scheme,
-  TApp,
-  TCon,
-  Subst,
-  Env,
-  Constraint,
-  Unifier,
-  equal,
-  tInt,
-  tBool,
-  print,
-  TUnion,
-  freeze,
-} from "./type";
+import { UnboundVariable } from "./errors";
+import { freeze, scheme, tBool, tInt } from "./type";
+import { Constraint, Env, Scheme, Subst, TCon, TVar, Type } from "./type";
 import { Binop, Expr } from "./syntax";
-import { snd, zip } from "./util";
-
-function assertUnreachable(x: never): never {
-  throw new Error("Didn't expect to get here");
-}
-
-const scheme = (qualifiers: readonly TVar[], type: Type): Scheme => {
-  if (type === undefined) {
-    throw new Error("scheme: type can't be undefined");
-  }
-  return {
-    tag: "Forall",
-    qualifiers,
-    type,
-  };
-};
-
-const isTCon = (t: Type): t is TCon => t.tag === "TCon";
-const isTVar = (t: Type): t is TVar => t.tag === "TVar";
-const isTApp = (t: Type): t is TApp => t.tag === "TApp";
-const isTUnion = (t: Type): t is TUnion => t.tag === "TUnion";
-const isScheme = (t: any): t is Scheme => t.tag === "Forall";
-
-function apply(s: Subst, type: Type): Type;
-function apply(s: Subst, scheme: Scheme): Scheme;
-function apply(s: Subst, types: readonly Type[]): readonly Type[];
-function apply(s: Subst, schemes: readonly Scheme[]): readonly Scheme[];
-function apply(s: Subst, constraint: Constraint): Constraint; // special case of Type[]
-function apply(s: Subst, constraint: readonly Constraint[]): readonly Constraint[]; // this should just work
-function apply(s: Subst, env: Env): Env;
-function apply(s: Subst, a: any): any {
-  // instance Substitutable Type
-  if (isTCon(a)) {
-    if (a.id && s.has(a.id)) {
-      return s.get(a.id);
-    }
-    const result: TCon = {
-      ...a,
-      params: apply(s, a.params),
-    };
-    return result;
-  }
-  if (isTVar(a)) {
-    return s.get(a.id) || a;
-  }
-  if (isTApp(a)) {
-    return {
-      ...a,
-      args: apply(s, a.args),
-      ret: apply(s, a.ret),
-    };
-  }
-  if (isTUnion(a)) {
-    return {
-      ...a,
-      types: apply(s, a.types),
-    };
-  }
-
-  // instance Substitutable Scheme
-  if (isScheme(a)) {
-    return scheme(
-      a.qualifiers,
-      apply(
-        // remove all TVars from the Substitution mapping that appear in the scheme as
-        // qualifiers.
-        // TODO: should this be using reduceRight to match Infer.hs' use of foldr?
-        a.qualifiers.reduceRight((accum, val) => accum.delete(val.id), s),
-        a.type
-      )
-    );
-  }
-
-  // instance Substitutable Constraint
-  // instance Substitutable a => Substitutable [a]
-  if (Array.isArray(a)) {
-    return a.map((t) => apply(s, t));
-  }
-
-  // instance Substitutable Env
-  if (Map.isMap(a)) {
-    return (a as Env).map((sc) => apply(s, sc));
-  }
-
-  throw new Error(`apply doesn't handle ${a}`);
-}
-
-function ftv(type: Type): Set<TVar>;
-function ftv(scheme: Scheme): Set<TVar>;
-function ftv(types: readonly Type[]): Set<TVar>;
-function ftv(schemes: readonly Scheme[]): Set<TVar>;
-function ftv(constraint: Constraint): Set<TVar>; // special case of Type[]
-function ftv(env: Env): Set<TVar>;
-function ftv(a: any): any {
-  // instance Substitutable Type
-  if (isTCon(a)) {
-    return Set.union(a.params.map(ftv));
-  }
-  if (isTVar(a)) {
-    return Set([a]); // Set.singleton a
-  }
-  if (isTApp(a)) {
-    return Set.union([...a.args.map(ftv), ftv(a.ret)]); // ftv t1 `Set.union` ftv t2
-  }
-  if (isTUnion(a)) {
-    return ftv(a.types);
-  }
-
-  // instance Substitutable Scheme
-  if (isScheme(a)) {
-    return ftv(a.type).subtract(a.qualifiers);
-  }
-
-  // instance Substitutable Constraint
-  // instance Substitutable a => Substitutable [a]
-  if (Array.isArray(a)) {
-    return Set.union(a.map(ftv));
-  }
-
-  // instance Substitutable Env
-  if (Map.isMap(a)) {
-    const env = a as Env;
-    return Set.union(env.valueSeq().map(ftv));
-  }
-
-  throw new Error(`ftv doesn't handle ${a}`);
-}
+import { zip, apply, ftv, assertUnreachable } from "./util";
+import { runSolve } from "./constraint-solver";
 
 type State = {
   count: number;
@@ -197,16 +52,6 @@ const closeOver = (t: Type): Scheme => {
   return result;
 };
 
-const lookup = (a: TVar, entries: readonly [TVar, TVar][]): TVar | null => {
-  for (const [k, v] of entries) {
-    // TODO: replace with IDs and generate names when printin
-    if (a.id === k.id) {
-      return v;
-    }
-  }
-  return null;
-};
-
 // remove duplicates from the array
 function nub(array: readonly TVar[]): readonly TVar[] {
   const ids: number[] = [];
@@ -219,9 +64,7 @@ function nub(array: readonly TVar[]): readonly TVar[] {
   });
 }
 
-// TODO: switch to IDs for types and then simplify the normalization by
-// generating names for type variables at print time.
-
+// TODO: simplify the normalization by generating names for type variables at print time.
 // Renames type variables so that they start with 'a' and there are no gaps
 const normalize = (sc: Scheme): Scheme => {
   // Returns the names of the free variables in a type
@@ -241,11 +84,11 @@ const normalize = (sc: Scheme): Scheme => {
   };
 
   const body = sc.type;
-  const keys = nub(fv(body));
+  const keys = nub(fv(body)).map((tv) => tv.id);
   const values: TVar[] = keys.map((key, index) => {
-    return {tag: "TVar", id: key.id, name: letters[index]};
-  })
-  const ord = zip(keys, values);
+    return { tag: "TVar", id: key, name: letterFromIndex(index) };
+  });
+  const mapping: Record<number, TVar> = Object.fromEntries(zip(keys, values));
 
   const normType = (type: Type): Type => {
     switch (type.tag) {
@@ -261,7 +104,7 @@ const normalize = (sc: Scheme): Scheme => {
       case "TCon":
         return type;
       case "TVar": {
-        const replacement = lookup(type, ord);
+        const replacement = mapping[type.id];
         if (replacement) {
           return replacement;
         } else {
@@ -274,12 +117,12 @@ const normalize = (sc: Scheme): Scheme => {
           types: type.types.map(normType),
         };
       }
-      default: 
-        assertUnreachable(type)
+      default:
+        assertUnreachable(type);
     }
   };
 
-  return scheme(ord.map(snd), normType(body));
+  return scheme(values, normType(body));
 };
 
 // Lookup type in the environment
@@ -294,42 +137,16 @@ const lookupEnv = (name: string, ctx: Context): Type => {
   return instantiate(value, ctx);
 };
 
-const letters = [
-  "a",
-  "b",
-  "c",
-  "d",
-  "e",
-  "f",
-  "g",
-  "h",
-  "i",
-  "j",
-  "k",
-  "l",
-  "m",
-  "n",
-  "o",
-  "p",
-  "q",
-  "r",
-  "s",
-  "t",
-  "u",
-  "v",
-  "w",
-  "x",
-  "y",
-  "z",
-];
+const letterFromIndex = (index: number): string =>
+  String.fromCharCode(97 + index);
 
-// TODO: use IDs for TVar's and defer naming until print time
+// TODO: defer naming until print time
 export const fresh = (ctx: Context): TVar => {
   ctx.state.count++;
   return {
     tag: "TVar",
     id: ctx.state.count,
-    name: letters[ctx.state.count],
+    name: letterFromIndex(ctx.state.count),
   };
 };
 
@@ -341,7 +158,7 @@ export const freshTCon = (ctx: Context, name: string): TCon => {
     name,
     params: [],
   };
-}
+};
 
 const instantiate = (sc: Scheme, ctx: Context): Type => {
   const freshQualifiers = sc.qualifiers.map(() => fresh(ctx));
@@ -358,36 +175,10 @@ const generalize = (env: Env, type: Type): Scheme => {
   return scheme(ftv(type).subtract(ftv(env)).toArray(), type);
 };
 
-const ops = (op: Binop): Type => {
-  switch (op) {
-    case "Add":
-      return {
-        tag: "TApp",
-        args: [tInt, tInt],
-        ret: tInt,
-      };
-    case "Mul":
-      return {
-        tag: "TApp",
-        args: [tInt, tInt],
-        ret: tInt,
-      };
-    case "Sub":
-      return {
-        tag: "TApp",
-        args: [tInt, tInt],
-        ret: tInt,
-      };
-    case "Eql":
-      return {
-        tag: "TApp",
-        args: [tInt, tInt],
-        ret: tBool,
-      };
-  }
-};
-
-const infer = (expr: Expr, ctx: Context): readonly [Type, readonly Constraint[]] => {
+const infer = (
+  expr: Expr,
+  ctx: Context
+): readonly [Type, readonly Constraint[]] => {
   switch (expr.tag) {
     case "Lit": {
       const lit = expr.value;
@@ -492,153 +283,31 @@ const infer = (expr: Expr, ctx: Context): readonly [Type, readonly Constraint[]]
   }
 };
 
-//
-// Constraint Solver
-//
-
-const emptySubst: Subst = Map();
-
-const runSolve = (cs: readonly Constraint[]): Subst => {
-  return solver([emptySubst, cs]);
-};
-
-const unifyMany = (ts1: readonly Type[], ts2: readonly Type[]): Subst => {
-  if (ts1.length !== ts2.length) {
-    throw new UnificationMismatch(ts1, ts2);
-  }
-  if (ts1.length === 0 && ts2.length === 0) {
-    return emptySubst;
-  }
-  const [t1, ...rest1] = ts1;
-  const [t2, ...rest2] = ts2;
-  const su1 = unifies(t1, t2);
-  const su2 = unifyMany(apply(su1, rest1), apply(su1, rest2));
-  return composeSubs(su2, su1);
-};
-
-export const unifies = (t1: Type, t2: Type): Subst => {
-  if (equal(t1, t2)) {
-    return emptySubst;
-  } else if (isTVar(t1)) {
-    return bind(t1, t2);
-  } else if (isTVar(t2)) {
-    return bind(t2, t1);
-  } else if (isTApp(t1) && isTApp(t2)) {
-    // infer() only ever creates a Lam node on the left side of a constraint
-    // and an App on the right side of a constraint so this check is sufficient.
-    if (t1.src === "Lam" && t2.src === "App") {
-      // partial application
-      if (t1.args.length > t2.args.length) {
-        const t1_partial: Type = {
-          tag: "TApp",
-          args: t1.args.slice(0, t2.args.length),
-          ret: {
-            tag: "TApp",
-            args: t1.args.slice(t2.args.length),
-            ret: t1.ret,
-          },
-          src: t1.src,
-        };
-        return unifyMany(
-          [...t1_partial.args, t1_partial.ret],
-          [...t2.args, t2.ret]
-        );
-      }
-
-      // subtyping: we ignore extra args
-      // TODO: Create a `isSubType` helper function
-      // TODO: update this once we support rest params
-      if (t1.args.length < t2.args.length) {
-        const t2_without_extra_args: Type = {
-          tag: "TApp",
-          args: t2.args.slice(0, t1.args.length),
-          ret: t2.ret,
-          src: t2.src,
-        }
-        return unifyMany(
-          [...t1.args, t1.ret],
-          [...t2_without_extra_args.args, t2_without_extra_args.ret]
-        );
-      }
-    }
-
-    // The reverse can happen when a callback is passed as an arg
-    if (t1.src === "App" && t2.src === "Lam") {
-      // Can partial application happen in this situation?
-
-      // subtyping: we ignore extra args
-      // TODO: Create a `isSubType` helper function
-      // TODO: update this once we support rest params
-      if (t1.args.length > t2.args.length) {
-        const t1_without_extra_args: Type = {
-          tag: "TApp",
-          args: t1.args.slice(0, t2.args.length),
-          ret: t1.ret,
-          src: t1.src,
-        }
-        return unifyMany(
-          [...t1_without_extra_args.args, t1_without_extra_args.ret],
-          [...t2.args, t2.ret]
-        );
-      }
-    }
-
-    // TODO: add support for optional params
-    // we can model optional params as union types, e.g. int | void
-
-    return unifyMany([...t1.args, t1.ret], [...t2.args, t2.ret]);
-  } else if (isTCon(t1) && isTCon(t2) && t1.name === t2.name) {
-    return unifyMany(t1.params, t2.params);
-  } else if (isTUnion(t1) && isTUnion(t2)) {
-    // Assume that the union types have been normalized by this point
-    // This only works if the types that make up the unions are ordered
-    // consistently.  Is there a way to do this?
-    return unifyMany(t1.types, t2.types);
-  } else {
-    // As long as the types haven't been frozen then this is okay
-    // NOTE: We may need to add .src info in the future if we notice
-    // any places where expected type widening is occurring.
-    if ("id" in t1 && "id" in t2 && !t1.frozen && !t2.frozen) {
-      const union: TUnion = {
-        tag: "TUnion",
-        types: [t1, t2],
+const ops = (op: Binop): Type => {
+  switch (op) {
+    case "Add":
+      return {
+        tag: "TApp",
+        args: [tInt, tInt],
+        ret: tInt,
       };
-      const result: Subst = Map([
-        [t1.id, union],
-        [t2.id, union],
-      ])
-      return result;
-    }
-
-    throw new UnificationFail(t1, t2);
+    case "Mul":
+      return {
+        tag: "TApp",
+        args: [tInt, tInt],
+        ret: tInt,
+      };
+    case "Sub":
+      return {
+        tag: "TApp",
+        args: [tInt, tInt],
+        ret: tInt,
+      };
+    case "Eql":
+      return {
+        tag: "TApp",
+        args: [tInt, tInt],
+        ret: tBool,
+      };
   }
-};
-
-const composeSubs = (s1: Subst, s2: Subst): Subst => {
-  return s2.map((t) => apply(s1, t)).merge(s1);
-};
-
-// Unification solver
-const solver = (u: Unifier): Subst => {
-  const [su, cs] = u;
-  if (cs.length === 0) {
-    return su;
-  }
-  const [[t1, t2], ...cs0] = cs;
-  const su1 = unifies(t1, t2);
-  return solver([composeSubs(su1, su), apply(su1, cs0)]);
-};
-
-const bind = (tv: TVar, t: Type): Subst => {
-  if (t.tag === "TVar" && t.id === tv.id) {
-    return emptySubst;
-  } else if (occursCheck(tv, t)) {
-    throw new InfiniteType(tv, t);
-  } else {
-    return Map([[tv.id, t]]);
-  }
-};
-
-const occursCheck = (tv: TVar, t: Type): boolean => {
-  return ftv(t).includes(tv);
 };
